+++ internal/component/prometheus/operator/configgen/config_gen_scrapeconfig_test.go
@@ -26,6 +26,7 @@ import (
 )
 
 func TestGenerateStaticScrapeConfigConfig(t *testing.T) {
+	HTTPS := "HTTPS"
 	suite := []struct {
 		name                   string
 		m                      *promopv1alpha1.ScrapeConfig
@@ -78,6 +79,56 @@ func TestGenerateStaticScrapeConfigConfig(t *testing.T) {
 						&targetgroup.Group{
 							Targets: []model.LabelSet{{model.AddressLabel: model.LabelValue("foo")}, {model.AddressLabel: model.LabelValue("bar")}},
 							Labels:  model.LabelSet{"foo": "bar"},
+							Source:  "scrapeConfig/operator/scrapeconfig/static/1",
+						},
+					},
+				},
+			},
+		},
+		{
+			name: "lowercase schema",
+			m: &promopv1alpha1.ScrapeConfig{
+				ObjectMeta: metav1.ObjectMeta{
+					Namespace: "operator",
+					Name:      "scrapeconfig",
+				},
+				Spec: promopv1alpha1.ScrapeConfigSpec{
+					Scheme: &HTTPS,
+				},
+			},
+			ep: promopv1alpha1.StaticConfig{
+				Targets: []promopv1alpha1.Target{"foo", "bar"},
+			},
+			expectedRelabels: util.Untab(`
+				- target_label: __meta_foo
+				  replacement: bar
+				- source_labels: [job]
+				  target_label: __tmp_prometheus_job_name
+				- replacement: operator
+				  target_label: __meta_kubernetes_scrapeconfig_namespace
+				- replacement: scrapeconfig
+				  target_label: __meta_kubernetes_scrapeconfig_name
+				- source_labels: [__address__]
+				  target_label: instance
+			`),
+			expected: &config.ScrapeConfig{
+				JobName:           "scrapeConfig/operator/scrapeconfig/static/1",
+				HonorTimestamps:   true,
+				ScrapeInterval:    model.Duration(time.Hour),
+				ScrapeTimeout:     model.Duration(42 * time.Second),
+				ScrapeProtocols:   config.DefaultScrapeProtocols,
+				EnableCompression: true,
+				MetricsPath:       "/metrics",
+				Scheme:            "https",
+				HTTPClientConfig: commonConfig.HTTPClientConfig{
+					FollowRedirects: true,
+					EnableHTTP2:     true,
+				},
+				ServiceDiscoveryConfigs: discovery.Configs{
+					discovery.StaticConfig{
+						&targetgroup.Group{
+							Targets: []model.LabelSet{{model.AddressLabel: model.LabelValue("foo")}, {model.AddressLabel: model.LabelValue("bar")}},
+							Labels:  model.LabelSet{},
 							Source:  "scrapeConfig/operator/scrapeconfig/static/1",
 						},
 					},